vc++2010fixes
Added a VC++ 2010 solution. The solution has three projects: fhe, helib,
and ntl. The ntl project is not provided, and you must build and add
this static library yourself as it is not a part of helib. Made a few
minor mods in order for the projects to compile properly. fhe is the
executable, and it executes Test_General.