FHEcontext is missing copy constructor and assignment operator.
Basically, FHEcontext holds a raw pointer to EncryptedArray.
We need to handle the deep copy of this raw pointer.
Without the deep copy, codes as follows fail.

```
FHEcontext context = pk.getContext(); //Bug here. A shadow copy of the Encrypted Array. So it will be deconstructed twice.
```

We can use a smart pointer rather than a raw one.
But considerring the backward compatibility, we can just add the copy constructor and assignment function.