binio: Performance: separate read_raw_int/write_raw_int to 32-bit/64-bit functions
Instead of relying on an intSize parameter, define two separate
functions, which makes profiling easier and is a first step towards
optimizing the implementation of each one.

Most of the code uses the 64-bit function, maybe it makes sense to
simply remove the 32-bit one which is only used when writing the
metadata of NTL vector.